Added a double-pulse test mode, more compile options, cleanup.
The double pulse test is a short test that allows us to see the
switching behaviour of the power stage. The 1st pulse lets current
build up in a stalled motor (or equivalent inductor) and then a
second pulse ends the test. The switch off of the 1st pulse and the
switch on of the second pulse are the relevant parts of the test,
it's where the built-up current is switched on-off.

Added macros to control inclusion of several feature in the compilation
(such as the ADC and the OCA), n order to allow code space for other
functionality such as the double-pulse test.