New attempt at measuring motor (driving) current.
Current change during the PWM period should be marginal (ripple),
typically <5%, should never be >10%. Therefore we can relax the
requirements and sample the current shunt at the beggining of the
ON period instead of at the end, thus making it so much easier.
Works decently, oscilates too much at very low currents (no load
on motor) and oscilates a little when the bridge is doing current
limiting, but I think it's reasonable now. It's reporting about
0.5-1A too much but that's perfectly fine.

The current shunt filter was changed, the resistor was reduced
from 2K2 to 680Ohm. The original values were low-passing too much,
attenuating the signal a lot at the beginning of the PWM period
(where we are now measuring).

Position of OCA and Isens where swapped in the schematics.