Bridge Ctrller sw: v26, security features (i2c write check, system master loss detection [still disabled]) and fast boot.
To reliability, the i2c register writes were changed to take 2 register
values, one the bitwise complement of the other. The bridge sw only updates
the register if so is true.

Code for detection of system master communication was added and compiles
but is disabled until tests can be done. This intents to detect loss of
master while the motor is running, and gracefully turn the motor off. From
time to time the sw will decrease the motor speed, unless it is "refreshed"
by a Speed register write.

The initial LED blink was removed to speed up bridge boot up. This also
fred some code space, allowing impementation of other more important
features such as the above ones.